requirements.txt
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
ci-info==0.2.0
click==8.1.2
configobj==5.0.6
configparser==5.2.0
cryptography==36.0.2
cycler==0.11.0
decorator==5.1.1
Elixir==0.7.1
et-xmlfile==1.1.0
etelemetry==0.3.0
filelock==3.6.0
fitz==0.0.1.dev2
fonttools==4.31.2
future==0.18.2
ghostscript==0.7
httplib2==0.20.4
idna==3.3
isodate==0.6.1
Jinja2==3.1.1
kiwisolver==1.4.2
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
networkx==2.7.1
nibabel==3.2.2
nipype==1.7.1
numpy==1.22.3
opencv-python==4.5.5.64
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
pathlib==1.0.1
pbr==5.8.1
pdfminer.six==20220319
pdftopng==0.2.3
Pillow==9.1.0
prov==2.0.0
pycparser==2.21
pydot==1.4.2
PyMuPDF==1.19.6
pyparsing==3.0.7
PyPDF2==1.26.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyxnat==1.4
rdflib==6.1.1
requests==2.27.1
scipy==1.8.0
simplejson==3.17.6
six==1.16.0
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.2
tabulate==0.8.9
Tempita==0.5.2
traits==6.3.2
urllib3==1.26.9
xlrd==0.7.1
xlwt==0.7.2
